+++ packages/core/src/features/home/feature.ts
@@ -3,5 +3,5 @@ import type { Feature } from "../../shared/feature.ts";
 export const home: Feature = {
   id: "home",
   dir: import.meta.url,
-  routes: () => [{ path: ":lang", file: "routes/home.tsx" }],
+  routes: () => [{ path: ":lang", file: "routes/home" }],
 };